refactor folder and adding session prolly
