+++ package.json
@@ -3,15 +3,17 @@
 	"version": "1.0.0",
 	"scripts": {
 		"test": "echo \"Error: no test specified\" && exit 1",
-		"start": "node server/server.js",
-		"dev": "nodemon server/server.js"
+		"start": "node index.js",
+		"dev": "nodemon index.js"
 	},
 	"author": "Muhammad Ridho & Arifin Ilham M.",
 	"license": "ISC",
 	"description": "Backend Area for Medix App",
 	"dependencies": {
 		"bcrypt": "^5.1.1",
+		"connect-flash": "^0.1.1",
 		"express": "^4.19.2",
+		"express-session": "^1.18.0",
 		"nanoid": "^3.3.7",
 		"nodemon": "^3.1.2"
 	}
